feat: update new

+++ flashlight/src/config.rs
@@ -32,7 +32,7 @@ pub(crate) struct FillerConfig {
 
     pub(crate) rebalance_settled_pnl_threshold: Option<u16>,
 
-    pub(crate) min_gas_balance_to_fill: Option<u16>,
+    pub(crate) min_gas_balance_to_fill: Option<f64>,
 }
 
 pub(crate) struct GlobalConfig {
